Move syntax tests in own file and enable test again
